elimine la clase profesor ademas elimine su vista, repositorio, servicio y entidad elimine la clase persona y agregue sus campos a la clase estudiante modifique las consultas de los repositorios q estaban vinculadas a persona modifique el data service y elimine todo lo relacionado con persona y profesor

+++ src/main/java/com/example/application/data/service/UserService.java
@@ -42,7 +42,7 @@ public class UserService {
         return repository.findAll(pageable);
     }
 
-    public int count() {
-        return (int) repository.count();
+    public long count() {
+        return repository.count();
     }
 }
